paddleaurum/requirements.txt
# ── Core Agent Frameworks ─────────────────────────────────────────────────────
langgraph>=0.1.19
langchain>=0.2.6
langchain-community>=0.2.6
crewai>=0.36.0
crewai-tools>=0.2.6

# ── LLM Backend (Local / Free) ────────────────────────────────────────────────
ollama>=0.2.1
langchain-ollama>=0.1.0

# ── Async HTTP & Networking ───────────────────────────────────────────────────
httpx>=0.27.0
aiohttp>=3.9.5

# ── Database ──────────────────────────────────────────────────────────────────
aiosqlite>=0.20.0

# ── Email ─────────────────────────────────────────────────────────────────────
aiosmtplib>=3.0.1

# ── Browser Automation ───────────────────────────────────────────────────────
playwright>=1.44.0

# ── Settings & Config ────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
PyYAML>=6.0.1

# ── Scheduling ───────────────────────────────────────────────────────────────
apscheduler>=3.10.4

# ── Search ───────────────────────────────────────────────────────────────────
duckduckgo-search>=6.1.9

# ── Data / ML ────────────────────────────────────────────────────────────────
numpy>=1.26.4

# ── Observability (Optional — LangSmith free tier) ───────────────────────────
langsmith>=0.1.77

# ── CLI / Logging ─────────────────────────────────────────────────────────────
rich>=13.7.1
typer>=0.12.3
